+++ src/types/buffer.template.h
@@ -19,8 +19,8 @@ typedef uint8_t $type$;
 
 typedef struct {
   $type$* data;
-  size_t count;
-  size_t capacity;
+  uint32_t count;
+  uint32_t capacity;
 } $name$Buffer;
 
 // Initialize a new buffer int instance.
